+++ packages/create/CHANGELOG.md
@@ -1,4 +1,4 @@
-# create-step-func-emailer
+# create-mailshot
 
 ## 0.2.2
 
@@ -16,4 +16,4 @@
 
 ### Minor Changes
 
-- [`72d23af`](https://github.com/mdwt/step-func-emailer/commit/72d23af8e07ab379c794d2918990f7a517a5fc85) Thanks [@mdwt](https://github.com/mdwt)! - Publish framework packages to npm and add create-step-func-emailer CLI
+- [`72d23af`](https://github.com/mdwt/mailshot/commit/72d23af8e07ab379c794d2918990f7a517a5fc85) Thanks [@mdwt](https://github.com/mdwt)! - Publish framework packages to npm and add create-mailshot CLI